config: Make GlobalConfiguration instance
Global configuration is accessed via a GlobalConfiguration instance.
The instance is conceptually a singleton, but singletons are not welcome
in libcamera so we must store the (preferably single) instance
somewhere.

This patch creates a GlobalConfiguration instance in CameraManager and
defines the corresponding access method.  CameraManager is typically
instantiated only once or a few times, it is accessible in
many places in libcamera and the configuration can be retrieved from it
and passed to other places if needed (it's read-only once created).
Using CameraManager for the purpose is still suboptimal and we use it
only due to lack of better options.  An alternative could be Logger,
which is still a singleton and it's accessible from everywhere.  But
with Logger, we have a chicken and egg problem -- GlobalConfiguration
should log contingent problems with the configuration when it's loaded
but if it is created in the logger then there are mutual infinite
recursive calls.  Perhaps some acceptable workaround could be found,
which would also allow an easier logging configuration.

If there are multiple CameraManager instances, there are also multiple
GlobalConfiguration instances.  They may or may not contain the same
data, depending on whether the global configuration file in the file
system was changed in the meantime.  This is dirty and a potential
source of trouble (although not expected to have much impact in
practice) but there is no idea how to do better if we want to avoid
having GlobalConfiguration singleton or dealing with the logging
problem.

The configuration is stored in CameraManager privately, it's not meant
to be accessed by applications.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>