libcamera: pipeline: Add R-Car Gen4 ISP pipeline
Add a pipeline handler for R-Car Gen4. The pipeline makes use of the
RkISP1 IPA and, depending on the kernel R-Car ISP driver, support the
same features as the RkISP1 pipeline handler.

The pipeline and IPA is tested with the Agc, Agw and
BlackLevelCorrection algorithms and produce a stable and good image.

Signed-off-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>