gstreamer: Adding property to control orientation
"orientation" parameter added to the libcamerasrc
element to control the orientation of the camera feed.

GST_PARAM_MUTABLE_READY allows the parameter to be changed
only when pipeline is in the READY state

The parameter is fed into the existing
CameraConfiguration orientation property

This allows users to specify the rotation of the video stream,
enhancing flexibility in camera applications.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>