config: Allow enabling software ISP in runtime
This patch allows enabling or disabling software ISP in runtime in
addition to compile time.  This can be useful for software ISP testing
on various platforms as well as for overriding the defaults in case the
defaults don't work well (e.g. hardware ISP may or may not work on
i.MX8MP depending on the kernel and libcamera patches present in the
given system).

The configuration is specified as follows:

  configuration:
    pipelines:
      simple:
        supported_devices:
        - driver: DRIVER-NAME
          software_isp: BOOLEAN
        - ...

For example:

  configuration:
    pipelines:
      simple:
        supported_devices:
        - driver: mxc-isi
          software_isp: true

The overall configuration of enabling or disabling software ISP may get
dropped in future but this patch is still useful in the meantime.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>