ipa: rkisp1: cproc: Report metadata
Presently the colour processing component exposes controls for
brightness, saturation, and contrast to the applications which are
handled and processed accordingly on the ISP.

The implementation lacks reporting the values that are set back to the
application.

Provide the inverse conversion functions to convert the values applied
and report them in the completed request metadata.

Signed-off-by: van Veen, Stephan <stephan.vanveen@karlstorz.com>
Signed-off-by: Zenker, Sebastian <sebastian.zenker@karlstorz.com>
[Kieran: Refactor commit message, fix checkstyle warnings]
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>