libcamera: device_enumerator: Use regex to match entity names
Some entities in a media graph have names that might differ from
implementation to implementation; for example the Camera Receiver
Unit and CSI-2 receiver on the KakiP board have entities with names
that include their address, in the form "csi-16000400.csi2". Passing
that entity name to DeviceMatch is too inflexible given it would only
work if that specific CSI-2 receiver were the one being used.

Update DeviceMatch.add() such that the string passed to it is used to
create a std::basic_regex, and use std::regex_search() instead of a
direct string comparison in DeviceMatch::match(). This allows us to
pass a string that will form a regex expression to match to the entity
instead, for example "csi-[0-9]{8}.csi2".

Signed-off-by: Daniel Scally <dan.scally@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>