libcamera: media_device: Use a regex to match entity name
Some kernel drivers give their entities names that will differ from
implementation to implementation; for example the drivers for the
Camera Receiver Unit and CSI-2 receiver in the RZ/V2H SoC give their
entities names that include their memory address, in the format
"csi-16000400.csi2". Passing that entity name to
MediaDevice::getEntityByName() is too inflexible given it would only
then work if that specific CSI-2 receiver were the one being used.

Update MediaDevice::getEntityByName() such that the string passed to
it is used to create a std::basic_regex, and use std::regex_search()
instead of a direct string comparison to find a matching entity. This
allows us to pass a string that will form a regex to match to the
entity instead, for example "csi-[0-9]{8}.csi2".

Signed-off-by: Daniel Scally <dan.scally@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>