libcamera: camera: Hook into the LayerManager
Add hooks into the CameraManager to call into the LayerManager on all
relevant public-facing interface of Camera.

The entry point for each function into the LayerManager has been
chosen based on the capabilities that we want to expose to Layer
implementations.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Isaac Scott <isaac.scott@ideasonboard.com>

---
Changes in v6:
- merge the metadata returned by the layers in requestCompleted into the
  request's metadata to forward it to the application

No change in v5

No change in v4

Changes in v3:
- send the copy of controls returned by LayerManager::start()
- each Camera now has a LayerController that it can call into,
  shortening each hook call significantly

Changes in v2:
- move the LayerManager out of Camera into the CameraManager
- remove hooks for generateConfiguration() and streams()
- Camera now passes itself into the layer hooks, as it is required for
  the LayerManager to fetch the closure based on the camera to pass to
  each layer