layer: Add layer to inject AeEnable control
Add a layer to implement the AeEnable control, so that we can remove it
from all IPAs and the Camera class, so that it is transparent to all
parties and it is available for applications to use.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>

---
For layers like this that we forsee to be "always built-in" we might
want a way to build it into libcamera instead of into separate shared
object files. In any case, this also serves as a demo of how I envision
a layer implementation to work.

Changes in v6:
- as metadata can no longer be set directly in the Request, instead
  return it in requestCompleted

No change in v5

No change in v4

No change in v3

Changes in v2:
- add init() and terminate()
- use closures
- remove layer namespacing
- add gnu_symbol_visibility