camera, ipa: all: Remove AeEnable handling
Remove AeEnable handling from the Camera and all IPAs, as it is now
handled by the inject_controls layer.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>

---
The mali-c55 IPA needs to support AnalogueGainMode and ExposureTimeMode,
but that is out of scope for this series.

No change in v6

No change in v5

No change in v4

No change in v3

No change in v2