libcamera: pipeline_handler: cancel waiting requests first
When stopping a camera, we may now find that there are requests
queued to the camera but not yet available in the pipeline handler.

These are "waitingReqeusts" which are not yet queued to the device.

While stopping a camera, we ask the pipeline to stop with stopDevice()
and then cancel any waiting requests after.

This however can lead to a race where having stopped the camera and
completed the requests, the pipeline may try to further consume requests
from the waiting lists.

Therefore empty the waitingRequests queue before stopping the device and
cancel the waitingRequests afterwards.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>