libcamera: simple: Consider raw output configurations
When generating simple pipeline configuration, raw output configurations
are generated but later filtered out.  Let's include processed and/or
raw output configurations depending on the requested stream roles.

Raw and processed formats are handled separately.  The intention is that
in case both raw and processed formats are requested then the raw
formats should be left intact to the extent possible and not be
influenced by the processed formats (implying that raw parameters
compatible with the processed output requirements must be requested by
the application).

Raw output configurations are identified by colorSpace being set to
ColorSpace::Raw.

This is another preparatory patch without making raw outputs working.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>