libcamera: layer_manager: Add LayerManager implementation
We want to be able to implement layers in libcamera, which conceptually
sit in between the Camera class and the application. This can be useful
for implementing things that don't belong inside the Camera/IPA nor inside
the application, such as intercepting and translation the AeEnable
control, or implementing the Sync algorithm.

To achieve this, first add a LayerManager implementation, which searches
for and loads layers from shared object files, and orchestrates
executing them. Actually calling into these functions from the Camera
class will be added in the following patch.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>