layer: Add layer that implements the sync algorithm
Implement the sync layer, which implements the sync algorithm. Any
Camera that supports the SyncAdjustment and FrameDurationLimits
controls, and that reports the SensorTimestamp metadata will
automatically be supported by this layer.

This code is heavily based on Raspberry Pi's sync algorithm
implementation, from the following files in commit [1]:
- src/ipa/rpi/controller/sync_algorithm.h
- src/ipa/rpi/controller/sync_status.h
- src/ipa/rpi/controller/rpi/sync.cpp
- src/ipa/rpi/controller/rpi/sync.h

[1] https://github.com/raspberrypi/libcamera/commit/d1a712060dcb0aab8564e0d1d86efe9ffcfee6b9

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>

---
Changes in v2:
- use UniqueFD for socket
- remove setting nonblock from fcntl() to socket()
  - add cloexec