pipeline: rkisp1: Properly handle the bufferCount set in the stream configuration
The bufferCount is reset to a hardcoded value of 4 in
RkISP1Path::validate().  Keep the default value of 4 but do not reset
it, if it was changed.  This allows the user to set bufferCount to an
arbitrary number of buffers which then can be allocated for example by
the FrameBufferAllocator. Internally maxQueuedRequestsDevice_ is used as
buffer count which is the maximum number of buffers queued to the device
at once.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>