libcamera: simple: Handle processed and raw formats separately
Let's handle both processed and/or raw output configurations, depending
on the requested stream roles.  In addition to the already handled
processed formats and sizes, this patch adds handling of raw formats and
sizes, which correspond to the capture formats and sizes.

Raw and processed formats are handled separately.  The intention is that
in case both raw and processed formats are requested then the raw
formats should be left intact to the extent possible and not be
influenced by the processed formats (implying that raw parameters
compatible with the processed output requirements must be requested by
the application).

Accompanying checks for invalid role specifications are introduced.

This is another preparatory patch without making raw outputs working.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>