libcamera: mali-c55: Use the RZG2LCRU class
Use the RZG2LCRU class to add the ability to the mali-c55 pipeline
handler to read frames from memory and process them. The data flow
is different to an inline configuration; instead of filling the
params buffer immediately on queueRequestDevice(), queued requests
are parked in a pending queue until a buffer from the CRU device is
available to be queued to that device. Once that buffer is filled
with image data by the CRU the parameter buffer is filled and the
image buffer queued to the ISP.

Signed-off-by: Daniel Scally <dan.scally@ideasonboard.com>