b4: Add .b4-config
Per the b4 documentation: B4 is a tool created to make it easier for
project developers and maintainers to use a distributed development
workflow that relies on patches and distribution lists for code
contributions and review.

B4 is now a common tool used by developers for Linux kernel workflows.

While the Linux kernel support is handled inside b4, other projects can
rely on a .b4-config file to achieve the same thing.

Add a .b4-config file to instruct b4 to send patches to the
libcamera-devel mailing list. This allows the command b4 prep
--auto-to-cc to automatically add the mailing to its To: list.

Command output without .b4-config :

  $ b4 prep --auto-to-cc
  Collecting To/Cc addresses
  No new addresses to add.

With .b4-config :

  $ b4 prep --auto-to-cc
  Collecting To/Cc addresses
      + To: libcamera-devel@lists.libcamera.org
  ---
  You can trim/expand this list with: b4 prep --edit-cover
  Invoking git-filter-repo to update the cover letter.
  New history written in 0.02 seconds...
  Completely finished after 0.04 seconds.

Signed-off-by: Benjamin Mugnier <benjamin.mugnier@foss.st.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>