libcamera: camera: Store `MetadataListPlan` in `Camera::Private`
Just like `ControlInfoMap controlInfo_` is a public member of the private
camera data for pipeline handlers to populate, add a `MetadataListPlan` as
well for the pipeline handlers to fill. This will be used to initialize
the `MetadataList` of each request created by the camera.

Also add `Camera::metadata()`, which makes it accessible for applications.

Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>