libipa: exposure_mode_helper: Take exposure/gain quantization into account
In ExposureModeHelper::splitExposure() the quantization of exposure time
and gain is not taken into account. This can lead to visible flicker
when the quantization steps are too big. As a preparation to fixing
that, add a function to set the sensor line length and the current
sensor mode helper and extend the clampXXX functions to return the
quantization error.

By default the exposure time quantization is assumed to be 1us and gain
is assumed to not be quantized at all.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Reviewed-by: Daniel Scally <dan.scally@ideasonboard.com>

---

Changes in v4:
- Improved documentation

Changes in v3:
- Collected tags
- Renamed lineLength to lineDuration
- Remove the "no functional changes" sentence, as the returned
  exposureTime/gain are now quantized, which happened outside of this
class before. So that is actually a functional change.