utils: codegen: proxy: Check `ipc_` instead of `isolate_`
Only try to send the "Exit" message if the `IPCPipeUnixSocket` object
exists. If the constructor fails, then `ipc_` might not be initialized,
which would lead to a nullptr dereference. This can still result in an
`ipc_->sendAsync()` call even if `ipc_` is unconnected, but that only
reports an error, and is still better than the status quo.

Bug: https://bugs.libcamera.org/show_bug.cgi?id=276
Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>