controls: Add SyncAdjustment and SyncInterface controls
Add a couple of controls related to the sync algorithm.

The SyncAdjustment control signals to the Camera the amount to adjust
the frame duration by for a single frame. Any Camera that implements
this control (and FrameDurationLimits and SensorTimestamp metadata) will
be supported by the sync algorithm layer.

The SyncInterface control signals to the sync layer which network
interface to use for listening (on client mode) or for multicasting (on
server mode). Given that enabling/disabling sync is done via control (as
opposed to always being on via environment variable), choosing the
network interface should be chosen in a similar manner.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Isaac Scott <isaac.scott@ideasonboard.com>

---
No change in v2