gstreamer: Report camera properties as device properties
Iterate over all libcamera camera properties and report them as
device properties. Each libcamera ControlType is mapped to the corresponding
gstreamer GType. If the ControlValue is an array of values
(ControlValue::isArray()), GValue with type GST_TYPE_ARRAY is used to set the
value of that ControlValue with the corresponding GType.

Signed-off-by: Umang Jain <uajain@igalia.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>