libcamera: media-device: Introduce MediaContext
Introduce the MediaContext class that represents a media device
context to which video devices and subdevices can be bound to.

A MediaContext can be created from a MediaDevice (which does not
support multi-context) and from SharedMediaDevice (which supports
multi-context).

Binding devices to a MediaContext created from MediaDevice is a nop,
while MediaContext created from SharedMediaDevice perform proper
bindings.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Reviewed-by: Isaac Scott <isaac.scott@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>