rpi/pisp: Set format and bind ISP subdev
The in-review patch series that introduces multi-context support
in the Linux kernel also modifies the RaspberryPi BE driver to
expose the ISP subdevice to userspace.

In order to have this libcamera series tested with the kernel
series, set format on the ISP subdev to pass link validation and
also bind the ISP subdev to the media device context.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>