Documentation: application: Update mediactl URL
The linux kernel documentation for the Media Controller has moved.
Update the URL accordingly to the new location.

The existing link pointed to the 'introduction' page - but this isn't
easy to identify or get an overview of the full documentation available
for media controller. Instead point the link to the top level of the
media controller userspace API pages.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>