libipa: Add SyncHelper
The sync layer allows any Camera to get sync support as long as it
implements:
- SyncAdjustment control
- FrameDurationLimits control
- SensorTimestamp metadata

While FrameDurationLimits is usually implemented by AGC and
SensorTimestamp is already implemented by all pipeline handlers, only
SyncAdjustment is left that needs to be implemented for all platforms.
It naturally would be a good idea to implement it in libipa so all IPAs
can easily be supported.

Add SyncHelper that wraps handling of the SyncAdjustment control.
Although it still needs to be plumbed into each IPA, it should mitigate
a lot of future code duplication.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>

---
Changes in v2:
- move functions to header
- make controlInfo static