ipa: rkisp1: agc: Add support for sync
Add sync support to the RkISP1 IPA using the SyncHelper from libipa. The
syncAdjustment is saved in the frameContext (as opposed to getting it
from the SyncHelper) to avoid potential races and ensure that the same
value that was set into vblank will be returned in metadata.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>

---
Changes in v2:
- move SyncHelper from inherited to a member of Agc
- cosmetic changes