libcamera: Rename ControlValue to ControlStorage
The ControlValue class in libcamera represents the storage location
of the values a Control represents.

When a ControlValue is constructed, assigned or set() the values
received as parameters are mem-copied inside the storage space allocated
inside the ControlValue.

This implies that creating ControlValue effectively mem-copies a
Control creating multiple storages from the same information.

To prepare to use wherever possible the newly introduced
ControlValueView in the libcamera API, rename the ControlValue
to ControlStorage.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>