libcamera: Make ControlValue a view
Rename ControlValueView to ControlValue in the whole codebase in
order to make the libcamera type ControlValue a read-only view over
control values stored somewhere else.

Usage of this view-like ControlValue class in public API allows to
provide the ControlList class with:

	const ControlValue get(unsigned int id) const;

that returns a reference to the value associated with a control id
without making a copy. Similarly, access to the min/max/def members of
a ControlInfo (as well as the list of supported values) now returns
a view without copying the stored values in a temporary ControlValue.

The actual control values can be copied in the caller by using the
already existing

	template<typename T> [[nodiscard]] auto get() const

function.

The rename on the whole code base has not required many manual
modifications in order to keep it compiling. The most notable ones are:

- ControlInfo::values() needs to re-create a vector in order to return
  it
- A new

	ControlList::set(unsigned int id, const ControlStorage &value)

  function had to be added to complement:

	void set(unsigned int id, const ControlValue &value);

  This new function serves to support direct initialization of
  controls in a control list with explicit values, in example:

  	lensCtrls.set(V4L2_CID_FOCUS_ABSOLUTE, static_cast<int32_t>(position));

  As a ControlValue cannot be constructed with a litterals or with
  values we need to go through a ControlStorage and then copy it to
  the control list backing storage.

- The ControlSerializer class functions binarySize() and store()
  have been moved to work with views

- The control_value.cpp test now actually tests ControlStorage

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>