libcamera: base: semaphore: Do not unlock prematurely
In `Semaphore::release()`, unlocking the mutex before signalling the condition
variable can be problematic, especially with "temporary" objects such as the
ones `BoundMethodBase::activatePack()` uses to handle `ConnectionTypeBlocking`.

Specifically, `Semaphore::acquire()` might lock the mutex after `Semaphore::release()`
has unlocked it, but before it had the chance to notify the condition variable.
In that case `Semaphore::acquire()` can succeed, and execution may proceed to
destroy the `Semaphore` object while the other thread is in the process of
running `std::condition_variable::notify_all()`.

[achernia@qti.qualcomm.com: Cherry-picked from:
  Repo: https://git.libcamera.org/libcamera/libcamera.git
  Branch: master
  Commit: 760971951420b4b31420cbea83adf227784e381d
]

Bug: https://bugs.libcamera.org/show_bug.cgi?id=225
Fixes: 66e7c5b774e288 ("libcamera: Add Semaphore class")
Change-Id: Ib02cb087ca413919ea702d5302413f2b3a4f2f66
Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>