meson: Check for Python modules manually
The meson python module's find_installation() method conveniently allows
testing if the interpreter provides a set of modules. We use it to check
for the presence of the modules required at build time. Unfortunately,
the meson python module is not meant to access the native Python
interpreter of the build machine, but the Python environment of the host
machine. Usage of find_installation() for this purpose is incorrect, it
may find modules in a different interpreter than the one used to run the
build scripts that use those modules.

Replace find_installation() with a manual check using python3 directly.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>