libcamera: media_device: Expand return values for populateEntities()
At the moment populateEntities() just returns true or false to indicate
success or failure. Although the populate() function assumes that the
topology is settled once the version number stops incrementing, that's
not guaranteed to be the case. One indicator that it's likely to increase
again in the future is the absence of interfaces for an entity, which
would only be created for V4L2 subdevices once the v4l2-async framework
had run to completion.

Update populateEntities() to return -EAGAIN in the event an entity has
no associated interface yet and pass that error out from populate to
indicate to callers that they should re-try the population at a later
time.

Signed-off-by: Daniel Scally <dan.scally@ideasonboard.com>