libcamera: software_isp: Reduce statistics image area
The statistics in software ISP is computed basically over the whole
image area, although only on part of the pixels.  It is not necessary to
cover the whole image area, it's sufficient to compute the statistics
let's say over the central area of 2/3 of the image width and height,
which should be both sufficient and faster.  The speedup is not that
important with the CPU implementation but it may save CPU work
more noticeably with GPU debayering implementation.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>