ipa: rkisp1: agc: awb: Apply digital gain using ISP
The digital gain is already calculated in the AGC algorithm but the
value is not consumed by the AWB algorithm, which is where the gain
stage is located. Add the needed plumbing to carry the value between the
two algorithms.

This is needed to get good images from sensors such as the IMX219 where
large sensor frames needs a small VBLANK value. This results in images
that are so under exposed that even the sensors analog gain can't
compensate. The digital gain stage of the ISP helps with this.

Signed-off-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>