py: gen-py-controls: Use generic workaround for enum prefixes
Detecting the prefix from the enum value strings doesn't work properly if
the individual values start with the same characters. This is the case, for
example, with On/Off and Start/Stop. Therefore, a hard-coded workaround was
required for the LensShadingMapMode (Off/On) control.

Since enum values typically use the control name as the prefix, a better
and more generic solution would be to use the name directly as the prefix.

However, since some existing control enums use a different prefix than
their name, the previous version is still required for these.

To eliminate the workaround and still remain compatible, the control name
is used as the prefix if the prefix determined with the original method
begins with the control name.

Signed-off-by: Matthias Fend <matthias.fend@emfend.at>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>